Enforce use of git if requested
If any git-related option is specified, including the newly-added
`--git` option, refuse to operate if the destination is not a bare
repo.

Note that specifying `--git` is not very useful right now, since you
need `--git-name` and `--git-email` anyway, but it will be useful once
sqlite2dir learns to pull git identity from git's config.